Fix level 6 bug in WebGL version
In the WebGL version, level 6 has a bug where the player after changing
color to light gray can pass through both white and light gray walls.
This is due to some WebAssembly error failing to create some collider on
them because the sprite texture is not marked as read/write, which is
apparently not necessary for the PC version nor the in-editor Play.